rhino: upgrade to 1.7r5; use release URL
While GitHub "archive" URLs are notorious for being rebuilt from
time to time and changing checksums, "release" URLs are static.

Upgrade to 1.7r5, which is the closest release which is available.

Return to using wget fetcher as it is faster than git and we do
not have a compelling reason to need the full clone.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>